# el-pasas/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "0.1",
    "title" : "eDocument API",
    "termsOfService" : ""
  },
  "host" : "el-pasas.azurewebsites.net",
  "basePath" : "/api",
...
  "security" : [ {
    "Token" : [ ]
  } ],
  "paths" : {
...
    "/personalIdentityCard/{userId}" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get Personal Identity Card for the User",
        "operationId" : "getPersonalIdentityCard",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card model",
            "schema" : {
              "$ref" : "#/definitions/PersonalIdentityCard"
            }
          },
          "404" : {
            "description" : "Personal Identity Card for this {userId} not found"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/image/small" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get small profile image for the Personal Identity Card which belongs to the User",
        "operationId" : "getPersonalIdentityCardImageSmall",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card small profile image"
          },
          "404" : {
            "description" : "Personal Identity Card small image for this {userId} not found"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/signature" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get signature image for the Personal Identity Card which belongs to the User",
        "operationId" : "getPersonalIdentityCardSignature",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card signature image"
          },
          "404" : {
            "description" : "Personal Identity Card signature image for this {userId} not found"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/qr" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get QR code for the Personal Identity Card which belongs to the User",
        "operationId" : "getPersonalIdentityCardQr",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card QR code"
          },
          "404" : {
            "description" : "Personal Identity Card QR code for this {userId} not found"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/nfc" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get NFC code for the Personal Identity Card which belongs to the User",
        "operationId" : "getPersonalIdentityCardNfc",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card NFC code"
          },
          "404" : {
            "description" : "Personal Identity Card NFC code for this {userId} not found"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/image" : {
      "get" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Get profile image for the Personal Identity Card which belongs to the User",
        "operationId" : "getPersonalIdentityCardImage",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Personal Identity Card profile image"
          },
          "404" : {
            "description" : "Personal Identity Card image for this {userId} not found"
          }
        }
      },
      "put" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Update profile image for the Personal Identity Card which belongs to the User",
        "operationId" : "putPersonalIdentityCardImage",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "image",
          "in" : "formData",
          "description" : "Image to upload",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully uploaded Personal Identity Card profile image"
          },
          "400" : {
            "description" : "Provided image is invalid"
          }
        }
      }
    },
    "/personalIdentityCard/{userId}/validate" : {
      "put" : {
        "tags" : [ "PersonalIdentityCard" ],
        "summary" : "Request to revalidate and extend expiration for Personal Identity Card",
        "operationId" : "putPersonalIdentityCardValidation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully requested Personal Identity Card revalidation"
          },
          "404" : {
            "description" : "Personal Identity Card for this {userId} not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Token" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "PersonalIdentityCard" : {
      "type" : "object",
      "properties" : {
        "surname" : {
          "type" : "string",
          "example" : "Bružaitė"
        },
        "givenName" : {
          "type" : "string",
          "example" : "Vigilija"
        },
        "nationality" : {
          "type" : "string",
          "example" : "Lietuvos Respublikos"
        },
        "sex" : {
          "type" : "string",
          "example" : "MOT/F"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "11 03 1978"
        },
        "personalNo" : {
          "type" : "integer",
          "example" : 47803111025
        },
        "cardNo" : {
          "type" : "integer",
          "example" : 56499778
        },
        "dateOfExpiry" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "04 07 2022"
        },
        "dateOfIssue" : {
          "type" : "string",
          "format" : "dd mm yyyy",
          "example" : "04 07 2012"
        },
        "authority" : {
          "type" : "string",
          "example" : "Vilnius (19)"
        },
        "footer" : {
          "type" : "string",
          "example" : "I<LTU56499778<047803111025<<<<7803118F2207043LTU<<<<<<<<<<<8BRUZAITE<<VIGILIJA<<<<<<<<<<<<",
          "minLength" : 90,
          "maxLength" : 90
        }
      },
      "example" : {
        "dateOfExpiry" : "04 07 2022",
        "nationality" : "Lietuvos Respublikos",
        "footer" : "I<LTU56499778<047803111025<<<<7803118F2207043LTU<<<<<<<<<<<8BRUZAITE<<VIGILIJA<<<<<<<<<<<<",
        "surname" : "Bružaitė",
        "givenName" : "Vigilija",
        "sex" : "MOT/F",
        "authority" : "Vilnius (19)",
        "personalNo" : 47803111025,
        "dateOfBirth" : "11 03 1978",
        "dateOfIssue" : "04 07 2012",
        "cardNo" : 56499778
      }
    },
...
  }
}